+++ src/composites.rs
@@ -100,10 +100,17 @@ type CompositeServiceVec<ReqBody, ResBody, Error> = Vec<(
     Box<dyn CompositedService<ReqBody, ResBody, Error> + Send>,
 )>;
 
-type CompositeMakeServiceVec<Target, ReqBody, ResBody, Error, MakeError> = Vec<(
+type CompositeMakeServiceVec<Target, ReqBody, ResBody, Error, MakeError> =
+    Vec<CompositeMakeServiceEntry<Target, ReqBody, ResBody, Error, MakeError>>;
+
+/// Service which can be composited with other services as part of a CompositeMakeService
+///
+/// Consists of a base path for requests which should be handled by this service, and a boxed
+/// MakeService.
+pub type CompositeMakeServiceEntry<Target, ReqBody, ResBody, Error, MakeError> = (
     &'static str,
     Box<dyn CompositedMakeService<Target, ReqBody, ResBody, Error, MakeError> + Send>,
-)>;
+);
 
 /// Wraps a vector of pairs, each consisting of a base path as a `&'static str`
 /// and a `MakeService` instance. Implements `Deref<Vec>` and `DerefMut<Vec>` so