+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Support conversion for AnyOf and OneOf
+- Expose CompositeMakeServiceEntry
 
 ### Changed